Added class TestExamineDialog, whose main purpose is to represent all questions in test.
TODO: write some logic for TestDialog:
			- timer;
			- check results;
			- calculate scores;

TODO: implement for Test(container class) some methods to save and load
tests. For now i'll use simple files to store data, because i dont know
sql(or other base). But i'll try to realise it in the way of the the
easiest reimplementaition to SQL in the future.